Move `broadcast` command to `chain` module
