+++ bin/aecli-chain.js
@@ -89,6 +89,19 @@ program
   .description('Real-time block monitoring')
   .action(async (...arguments) => await Chain.play(utils.cli.getCmdFromArguments(arguments)))
 
+// ## Initialize `broadcast` command
+//
+// You can use this command to send `transaction` to the `chain`
+//
+// Example: `aecli tx spend ak_2a1j2Mk9YSmC1gioUq4PWRm3bsv887MbuRVwyv4KaUGoR1eiKi ak_AgV756Vfo99juwzNVgnjP1gXX1op1QN3NXTxvkPnHJPUDE8NT 100`
+program
+  .command('broadcast <tx>')
+  .option('--waitMined', 'Transaction payload.')
+  .option('--verify', 'Verify Transaction before broadcast.')
+  .description('Send transaction to the chain')
+  .action(async (tx, ...arguments) => await Chain.broadcast(tx, utils.cli.getCmdFromArguments(arguments)))
+
+
 // Handle unknown command's
 program.on('command:*', () => utils.errors.unknownCommandHandler(program)())
